app/requirements.txt
flask
gunicorn
psycopg2-binary==2.9.9
redis
python-dotenv
